Replace Both class by a simple dispatcher to #both
