+++ index.es.js
@@ -11,7 +11,6 @@ export {
 export * from './src/dispatchers';
 
 export {after} from './src/after';
-export {both} from './src/both';
 export {cache} from './src/cache';
 export {chainRec} from './src/chain-rec';
 export {encase, encase2, encase3, attempt, attempt as try} from './src/encase';
